Fixed relations ownership of Product and Dish. Changed Product.amount from double to int. Added increase/decrease stock of Product methods. Modified updateProduct and updateDish.

+++ src/main/java/com/backend/tokokantjil/dtos/inputs/ProductInputDto.java
@@ -12,7 +12,7 @@ public class ProductInputDto {
     @NotBlank
     public String name;
     @Positive
-    public double amount;
+    public int amount;
     @NotBlank
     public String amountUnit;
     @PositiveOrZero
