+++ src/main/java/com/backend/tokokantjil/dtos/outputs/ProductOutputDto.java
@@ -6,7 +6,7 @@ public class ProductOutputDto {
     private Long id;
     private String name;
     private State state;
-    private double amount;
+    private int amount;
     private String amountUnit;
     private double buyPrice;
     private double sellPrice;
@@ -38,11 +38,11 @@ public class ProductOutputDto {
         this.state = state;
     }
 
-    public double getAmount() {
+    public int getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(int amount) {
         this.amount = amount;
     }
